requirements.txt
flask
omxplayer-wrapper
werkzeug
flask-cors
flask-socketio
eventlet
pillow
netifaces
